CloudRun/requirements.txt
apache-airflow-providers-google==6.8.0
google-cloud-pubsub
sqlalchemy-spanner==<compatible-version>
